Added option to use existing segmentation
It is now possible to add a TextGrid file with vowel boundaries. This segmentation will be used to select the vowel segments instead of the syllable based automatic vowel segmentation.

The central 2/3rds of each vowel segment will be used (according to the list of vowel symbols).

If a phoneme tier number with vowel symbols is entered, the corresponding TextGrid file is opened. If no TextGrid file with the same name as the audio file is found, the user is asked to select the file.

If no space separated vowel symbols are given, the default SAMPA list for the language is used (Pinyin for ZH).

In automatic mode, TextGrids with wildcard * wil be treated as the corresponding audio files.

Support for all PRaat audio formats has been added.